requirements.txt
fastapi
uvicorn[standard]
pandas
selenium
python-dotenv
supabase
pyjwt
webdriver-manager
pyppeteer==2.0.0
httpx
apscheduler